Firmware: Got basic card reading working!
This version is a bit rough and ready, but it does work. The basic idea is that it
reads the card, and lights up green for a Mifare-ish card, or blue for a FeliCa card
(e.g Octopus)

The intention is to rewrite this to be more state driven, to make it easier to handle.