Hardware V1.1 - 3.3V with chunkier 5V regulator
The 1.0 5V regulator wasn't chunky enough, so this board has a larger
one to support the current draw of the PN532 module. Additionally, a
capacitor has been added near to the PN532 module connector to assist in
smoothing the supply to it.

Most of the board now runs at 3.3V, derived from the 5V rail. This is to
support direct connection to an ACNode, which only support 3.3V IO. This
required slowing the AVR down to 12MHz. An alternative option would have
been to use a level shifter on the TX line.